Finish stats function (#9)
* Add function for bot stats

* Refactor code to separate files for feedback and broadcast

* Add feedback for OHS

* Refine and finish stats function

* Remove useless log print in echo

* Better responses for stats function
